+++ mojo/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.0.48"
+__version__ = "1.0.49"
 
 from mojo.helpers.response import JsonResponse